fix: ホームの内容を本棚に移行
