+++ native/app/containers/index.ts
@@ -9,4 +9,4 @@ export { default as ContactEdit } from './ContactEdit';
 export { default as ProfileEdit } from './ProfileEdit';
 export { default as OwnProfile } from './OwnProfile';
 export { default as SignInSelect } from './SignInSelect';
-export { default as Home } from './Home';
+export { default as Bookshelf } from './Bookshelf';
